Aplicado color a las estrellas de fragment_resena.xml
Hecha la funcion para guardar un obj de tipo Resena en el que se
almacena:
Fecha de la resena
contenido
nota
falta por hacer que recoja el usuario que la usa.

+++ app/src/main/java/com/example/travelwithmeapp/models/Resena.kt
@@ -7,7 +7,7 @@ import java.time.OffsetDateTime
 
 data class Resena @RequiresApi(Build.VERSION_CODES.O) constructor(
     var id: Long = -1,
-    val idUsuario: Long = -1,
+    val idUsuario: Int = -1,
     val fecha: OffsetDateTime = OffsetDateTime.now(),
     val titulo: String = "",
     val contenido: String = "",
